Reorganize the logical plan related code in proto to be consistent with the physical plan code
